+++ freeswitch-esl-spring-boot-starter/src/main/java/link/thingscloud/freeswitch/esl/spring/boot/starter/properties/ServerProperties.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package link.thingscloud.freeswitch.esl.spring.boot.starter.propeties;
+package link.thingscloud.freeswitch.esl.spring.boot.starter.properties;
 
 import lombok.Data;
